Adds ScalableRect component to the codebase...
and updates Blockchain Learn page to use it.

Why? I had to add a few videos to Blockchain Learn page in a responsive
way, so that they keep their side ratio, while being scaled according to
the current viewport size, and, though it is a well-known HTML/CSS
trick, very soon I found out that it takes too much time to implement in
different places from scratch, without making mistakes on the way; thus
this new component, which takes care about all the magic, allowing just
to specify which sides ratio we need to render a block.