Yet another snapshot test fixed
NOTE: This snapshot turns to be differenly generated with Node v8.9.2
(currently used by the codebase) and Node v8.9.3 (the latest I have on
my machine). At the same time, the latest Node LTS is v8.9.4. It should
be carefully double-checked, whether it is safe to update to Node
v8.9.4, and if there are any problems, they should be fixed.